+++ backend/app/models/gasto.py
@@ -1,6 +1,7 @@
 from sqlalchemy import BigInteger, String, Numeric, Text, Date, DateTime, ForeignKey, CheckConstraint, text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from datetime import datetime, date
+from decimal import Decimal
 from app.core.database import Base
 
 
@@ -15,7 +16,7 @@ class Gasto(Base):
     categoria_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("categorias_gasto.id"), nullable=False)
     lote_id: Mapped[int | None] = mapped_column(BigInteger, ForeignKey("lotes.id"), nullable=True)
     descripcion: Mapped[str] = mapped_column(String(250), nullable=False)
-    valor: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False)
+    valor: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False)
     proveedor: Mapped[str | None] = mapped_column(String(150), nullable=True)
     observaciones: Mapped[str | None] = mapped_column(Text, nullable=True)
     registrado_por: Mapped[int] = mapped_column(BigInteger, ForeignKey("usuarios.id"), nullable=False)
