+++ backend/app/models/venta.py
@@ -1,6 +1,7 @@
 from sqlalchemy import BigInteger, String, Numeric, Text, Date, DateTime, ForeignKey, CheckConstraint, text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from datetime import datetime, date
+from decimal import Decimal
 from app.core.database import Base
 
 
@@ -13,7 +14,7 @@ class Venta(Base):
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     fecha: Mapped[date] = mapped_column(Date, nullable=False)
     cliente: Mapped[str | None] = mapped_column(String(150), nullable=True)
-    total: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False, server_default=text("0"))
+    total: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False, server_default=text("0"))
     observaciones: Mapped[str | None] = mapped_column(Text, nullable=True)
     registrado_por: Mapped[int] = mapped_column(BigInteger, ForeignKey("usuarios.id"), nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, server_default=text("NOW()"))
@@ -32,9 +33,9 @@ class DetalleVenta(Base):
 
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     venta_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("ventas.id", ondelete="CASCADE"), nullable=False)
-    cantidad: Mapped[float] = mapped_column(Numeric(12, 3), nullable=False)
-    precio_unitario: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False)
-    subtotal: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False)
+    cantidad: Mapped[Decimal] = mapped_column(Numeric(12, 3), nullable=False)
+    precio_unitario: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False)
+    subtotal: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False)
     lote_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("lotes.id"), nullable=False)
 
     venta = relationship("Venta", back_populates="detalles")
